pyproject.toml
[project]
name = "spacesim2"
version = "0.2.0"
description = "A turn-based economic simulation modeling interplanetary trade"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "pygame>=2.5.0; platform_system != 'Windows' or python_version < '3.12'",
    "pyyaml>=6.0.2",
    "scipy>=1.15.2",
    "typing-extensions>=4.15.0",
]

[project.scripts]
spacesim2 = "spacesim2.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
    "black>=23.0.0",
    "rich>=13.0.0",
]
analysis = [
    "polars>=0.20.0", # Fast DataFrame library
    "pyarrow>=14.0.0", # Parquet support
    "plotly>=5.18.0", # Interactive visualizations
    "marimo>=0.1.0", # Modern reactive notebook
    "tqdm>=4.66.0", # Progress bars
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["spacesim2*"]
exclude = ["tests*", "temp*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
